+++ dev/sparktestsupport/modules.py
@@ -509,8 +509,11 @@ pyspark_sql = Module(
         "pyspark.sql.tests.test_conf",
         "pyspark.sql.tests.test_context",
         "pyspark.sql.tests.test_dataframe",
+        "pyspark.sql.tests.test_collection",
+        "pyspark.sql.tests.test_creation",
         "pyspark.sql.tests.test_listener",
         "pyspark.sql.tests.test_observation",
+        "pyspark.sql.tests.test_repartition",
         "pyspark.sql.tests.test_stat",
         "pyspark.sql.tests.test_datasources",
         "pyspark.sql.tests.test_errors",
@@ -1022,7 +1025,10 @@ pyspark_connect = Module(
         "pyspark.sql.tests.connect.test_parity_functions",
         "pyspark.sql.tests.connect.test_parity_group",
         "pyspark.sql.tests.connect.test_parity_dataframe",
+        "pyspark.sql.tests.connect.test_parity_collection",
+        "pyspark.sql.tests.connect.test_parity_creation",
         "pyspark.sql.tests.connect.test_parity_observation",
+        "pyspark.sql.tests.connect.test_parity_repartition",
         "pyspark.sql.tests.connect.test_parity_stat",
         "pyspark.sql.tests.connect.test_parity_types",
         "pyspark.sql.tests.connect.test_parity_column",
