+++ python/pyspark/sql/tests/connect/test_parity_collection.py
@@ -15,32 +15,40 @@
 # limitations under the License.
 #
 
-import unittest
-
-from pyspark.sql.tests.test_dataframe import DataFrameTestsMixin
+from pyspark.sql.tests.test_collection import DataFrameCollectionTestsMixin
 from pyspark.testing.connectutils import ReusedConnectTestCase
 
 
-class DataFrameParityTests(DataFrameTestsMixin, ReusedConnectTestCase):
-    @unittest.skip("Spark Connect does not support RDD but the tests depend on them.")
-    def test_help_command(self):
-        super().test_help_command()
+class DataFrameCollectionParityTests(
+    DataFrameCollectionTestsMixin,
+    ReusedConnectTestCase,
+):
+    def test_to_local_iterator_not_fully_consumed(self):
+        self.check_to_local_iterator_not_fully_consumed()
 
-    # Spark Connect throws `IllegalArgumentException` when calling `collect` instead of `sample`.
-    def test_sample(self):
-        super().test_sample()
+    def test_to_pandas_for_array_of_struct(self):
+        # Spark Connect's implementation is based on Arrow.
+        super().check_to_pandas_for_array_of_struct(True)
 
-    @unittest.skip("Spark Connect does not support RDD but the tests depend on them.")
-    def test_toDF_with_schema_string(self):
-        super().test_toDF_with_schema_string()
+    def test_to_pandas_from_null_dataframe(self):
+        self.check_to_pandas_from_null_dataframe()
 
-    def test_toDF_with_string(self):
-        super().test_toDF_with_string()
+    def test_to_pandas_on_cross_join(self):
+        self.check_to_pandas_on_cross_join()
+
+    def test_to_pandas_from_empty_dataframe(self):
+        self.check_to_pandas_from_empty_dataframe()
+
+    def test_to_pandas_with_duplicated_column_names(self):
+        self.check_to_pandas_with_duplicated_column_names()
+
+    def test_to_pandas_from_mixed_dataframe(self):
+        self.check_to_pandas_from_mixed_dataframe()
 
 
 if __name__ == "__main__":
     import unittest
-    from pyspark.sql.tests.connect.test_parity_dataframe import *  # noqa: F401
+    from pyspark.sql.tests.connect.test_parity_collection import *  # noqa: F401
 
     try:
         import xmlrunner  # type: ignore[import]
